add api-version line
This is needed to prevent CratesPlus from operating in compatibility mode. 
All item names must use new material names before this can be accepted. 

Spigot's built-in compatibility transformers are not efficient and should not be used.  As CP 5.x is intended for 1.14.x+ this line should be added.  If CP uses old item names, this WILL cause the plugin to fail.